+++ GlobalPayments-iOS-SDK/Classes/Mapping/GpApiMapping.swift
@@ -35,6 +35,9 @@ public struct GpApiMapping {
                 transaction.cardLast4 = card.getValue(key: "masked_number_last4")
                 transaction.cardType = card.getValue(key: "brand")
                 transaction.cvnResponseMessage = card.getValue(key: "cvv_result")
+                transaction.avsResponseCode = card.getValue(key: "avs_postal_code_result") ?? .empty
+                transaction.avsAddressResponse = card.getValue(key: "avs_address_result") ?? .empty
+                transaction.avsResponseMessage = card.getValue(key: "avs_action") ?? .empty
             }
         }