+++ GlobalPayments-iOS-SDK/Classes/PaymentMethods/Credit.swift
@@ -80,15 +80,15 @@ public class Credit: NSObject, PaymentMethod, Encryptable, Tokenizable, Chargeab
         return AuthorizationBuilder(transactionType: .verify, paymentMethod: self)
     }
 
-    public func tokenize(configName: String = "default", completion: ((String?, Error?) -> Void)?) {
-        tokenize(validateCard: true, configName: configName, completion: completion)
+    public func tokenize(configName: String = "default", paymentMethodUsageMode: PaymentMethodUsageMode = .multiple, completion: ((String?, Error?) -> Void)?) {
+        tokenize(validateCard: true, configName: configName,paymentMethodUsageMode: paymentMethodUsageMode, completion: completion)
     }
 
-    public func tokenize(validateCard: Bool, configName: String = "default", completion: ((String?, Error?) -> Void)?) {
+    public func tokenize(validateCard: Bool, configName: String = "default", paymentMethodUsageMode: PaymentMethodUsageMode = .multiple, completion: ((String?, Error?) -> Void)?) {
         let type: TransactionType = validateCard ? .verify : .tokenize
         AuthorizationBuilder(transactionType: type, paymentMethod: self)
             .withRequestMultiUseToken(validateCard)
-            .withTokenUsageMode(.multiple)
+            .withPaymentMethodUsageMode(paymentMethodUsageMode)
             .execute(configName: configName, completion: { transaction, error in
                 completion?(transaction?.token, error)
             })